instrumentation/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "App using FastAPI LASER"
authors = ["Tech <tech@settle.eu>"]
license = "Private"
packages = [
    { include = "app", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi-laser="0.1.0"
gunicorn = "^20.0.4"
fastapi = {version = "^0.88.0", extras = ["all"]}
loguru = "^0.5.3"
SQLAlchemy = "^1.4.22"
alembic = "^1.5.1"
psycopg2-binary = "^2.8.6"
google-cloud-secret-manager = "^2.1.0"
google-cloud-logging = "^2.1.1"
databases = {version = "0.4.2", extras = ["postgresql"]}
websockets = "^10.1"
httpx = "^0.23.1"
typer = "^0.3.2"
google-auth = "^2.6.2"
redis = "^4.3.4"
tenacity = "^8.1.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
# https://python-poetry.org/docs/dependency-specification
# TEST: Uncomment to include the current app in editable install mode
flake8 = "^3.8.4"
pylint = "^2.6.0"
black = "^21.12b0"
isort = "^5.6.4"
autoflake = "^1.4"
pytest = "^6.2.1"
pytest-cov = "^2.10.1"
pytest-xdist = "^2.2.0"
pytest-asyncio = "^0.14.0"
pytest-mock = "^3.7.0"
pytest-recording = "^0.11.0"
freezegun = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
db-migrations = "app.database.migrations.cli:cli"
probe-db = "app.probes:db_health_check"
